Module_3/stock_dashboard/requirements.txt
streamlit==1.38.0
yfinance==0.2.44
pandas==2.2.3
plotly==5.24.1
transformers==4.45.2
torch==2.5.0
python-dotenv==1.0.1
requests==2.31.0
